@@ -14,6 +14,8 @@ class DocumentListScreen extends ConsumerStatefulWidget {
   ConsumerState<DocumentListScreen> createState() => _DocumentListScreenState();
 }
 
+enum SortOption { date, name, type }
+
 class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
     with SingleTickerProviderStateMixin {
   late TabController _tabController;
@@ -23,12 +25,16 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
   bool _isGridView = true;
   DocumentType? _selectedType;
   bool _isLoadingMore = false;
+  SortOption _currentSort = SortOption.date;
+  bool _sortDescending = true;
 
   @override
   void initState() {
     super.initState();
-    _tabController =
-        TabController(length: DocumentType.values.length + 1, vsync: this);
+    _tabController = TabController(
+      length: DocumentType.values.length + 1,
+      vsync: this,
+    );
     _scrollController.addListener(_onScroll);
   }
 
@@ -89,17 +95,34 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
           ),
           PopupMenuButton<String>(
             onSelected: (value) {
-              switch (value) {
-                case 'sort_date':
-                  // TODO: Implement sorting
-                  break;
-                case 'sort_name':
-                  // TODO: Implement sorting
-                  break;
-                case 'sort_type':
-                  // TODO: Implement sorting
-                  break;
-              }
+              setState(() {
+                switch (value) {
+                  case 'sort_date':
+                    if (_currentSort == SortOption.date) {
+                      _sortDescending = !_sortDescending;
+                    } else {
+                      _currentSort = SortOption.date;
+                      _sortDescending = true;
+                    }
+                    break;
+                  case 'sort_name':
+                    if (_currentSort == SortOption.name) {
+                      _sortDescending = !_sortDescending;
+                    } else {
+                      _currentSort = SortOption.name;
+                      _sortDescending = false;
+                    }
+                    break;
+                  case 'sort_type':
+                    if (_currentSort == SortOption.type) {
+                      _sortDescending = !_sortDescending;
+                    } else {
+                      _currentSort = SortOption.type;
+                      _sortDescending = false;
+                    }
+                    break;
+                }
+              });
             },
             itemBuilder: (context) => [
               const PopupMenuItem(
@@ -140,8 +163,10 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
           child: Column(
             children: [
               Padding(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 16,
+                  vertical: 8,
+                ),
                 child: TextField(
                   controller: _searchController,
                   decoration: InputDecoration(
@@ -163,7 +188,7 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
                       borderSide: BorderSide.none,
                     ),
                     filled: true,
-                    fillColor: Colors.white.withOpacity(0.9),
+                    fillColor: Colors.white.withValues(alpha: 0.9),
                   ),
                   onChanged: (value) {
                     setState(() {
@@ -188,31 +213,36 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
                 isScrollable: true,
                 indicatorColor: Colors.white,
                 labelColor: Colors.white,
-                unselectedLabelColor: Colors.white.withOpacity(0.7),
+                unselectedLabelColor: Colors.white.withValues(alpha: 0.7),
                 onTap: (index) {
-                  final newType =
-                      index == 0 ? null : DocumentType.values[index - 1];
+                  final newType = index == 0
+                      ? null
+                      : DocumentType.values[index - 1];
                   setState(() {
                     _selectedType = newType;
                   });
                   // Filter documents by type
-                  ref.read(documentNotifierProvider.notifier).filterDocuments(
+                  ref
+                      .read(documentNotifierProvider.notifier)
+                      .filterDocuments(
                         type: newType,
                         searchQuery: _searchQuery.isEmpty ? null : _searchQuery,
                       );
                 },
                 tabs: [
                   const Tab(text: 'All'),
-                  ...DocumentType.values.map((type) => Tab(
-                        child: Row(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Icon(_getDocumentTypeIcon(type), size: 16),
-                            const SizedBox(width: 4),
-                            Text(type.displayName),
-                          ],
-                        ),
-                      )),
+                  ...DocumentType.values.map(
+                    (type) => Tab(
+                      child: Row(
+                        mainAxisSize: MainAxisSize.min,
+                        children: [
+                          Icon(_getDocumentTypeIcon(type), size: 16),
+                          const SizedBox(width: 4),
+                          Text(type.displayName),
+                        ],
+                      ),
+                    ),
+                  ),
                 ],
               ),
             ],
@@ -232,11 +262,49 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
     );
   }
 
+  List<Document> _sortDocuments(List<Document> documents) {
+    final sorted = List<Document>.from(documents);
+
+    switch (_currentSort) {
+      case SortOption.date:
+        sorted.sort(
+          (a, b) => _sortDescending
+              ? b.scanDate.compareTo(a.scanDate)
+              : a.scanDate.compareTo(b.scanDate),
+        );
+        break;
+      case SortOption.name:
+        sorted.sort(
+          (a, b) => _sortDescending
+              ? b.title.toLowerCase().compareTo(a.title.toLowerCase())
+              : a.title.toLowerCase().compareTo(b.title.toLowerCase()),
+        );
+        break;
+      case SortOption.type:
+        sorted.sort((a, b) {
+          final comparison = _sortDescending
+              ? b.type.name.compareTo(a.type.name)
+              : a.type.name.compareTo(b.type.name);
+          // If types are the same, sort by date as secondary
+          if (comparison == 0) {
+            return b.scanDate.compareTo(a.scanDate);
+          }
+          return comparison;
+        });
+        break;
+    }
+
+    return sorted;
+  }
+
   Widget _buildDocumentList(List<Document> documents) {
     // Documents are already filtered by database query
     if (documents.isEmpty) {
       return _buildEmptyState();
     }
+
+    // Apply sorting
+    final sortedDocuments = _sortDocuments(documents);
 
     final notifier = ref.read(documentNotifierProvider.notifier);
     final hasMore = notifier.hasMore;
@@ -253,7 +321,7 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
           return false;
         },
         child: DocumentGrid(
-          documents: documents,
+          documents: sortedDocuments,
           onDocumentTap: _navigateToDocumentDetail,
         ),
       );
@@ -261,9 +329,9 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
       return ListView.builder(
         controller: _scrollController,
         padding: const EdgeInsets.all(16),
-        itemCount: documents.length + (hasMore ? 1 : 0),
+        itemCount: sortedDocuments.length + (hasMore ? 1 : 0),
         itemBuilder: (context, index) {
-          if (index >= documents.length) {
+          if (index >= sortedDocuments.length) {
             // Loading more indicator
             return const Center(
               child: Padding(
@@ -272,7 +340,7 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
               ),
             );
           }
-          final document = documents[index];
+          final document = sortedDocuments[index];
           return DocumentListItem(
             document: document,
             onTap: () => _navigateToDocumentDetail(document),
@@ -291,7 +359,7 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
           Icon(
             Icons.folder_open,
             size: 64,
-            color: Theme.of(context).colorScheme.primary.withOpacity(0.5),
+            color: Theme.of(context).colorScheme.primary.withValues(alpha: 0.5),
           ),
           const SizedBox(height: 16),
           Text(
@@ -304,9 +372,10 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
                 ? 'Try adjusting your search terms'
                 : 'Start by scanning your first document',
             style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                  color:
-                      Theme.of(context).colorScheme.onSurface.withOpacity(0.6),
-                ),
+              color: Theme.of(
+                context,
+              ).colorScheme.onSurface.withValues(alpha: 0.6),
+            ),
           ),
           if (_searchQuery.isEmpty) ...[
             const SizedBox(height: 24),
@@ -337,9 +406,9 @@ class _DocumentListScreenState extends ConsumerState<DocumentListScreen>
           const SizedBox(height: 8),
           Text(
             error,
-            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                  color: Colors.red,
-                ),
+            style: Theme.of(
+              context,
+            ).textTheme.bodyMedium?.copyWith(color: Colors.red),
             textAlign: TextAlign.center,
           ),
           const SizedBox(height: 24),
