@@ -9,7 +9,6 @@ import 'providers/troubleshooting_logger_provider.dart';
 import 'services/database_service.dart';
 import 'services/encryption_service.dart';
 import 'services/ocr_service.dart';
-import 'services/camera_service.dart';
 import 'services/storage_provider_service.dart';
 import 'utils/navigation_observer.dart';
 import 'utils/error_handler.dart';
@@ -49,15 +48,13 @@ class OCRixApp extends ConsumerWidget {
       navigatorObservers: [navigationObserver],
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(
-          seedColor:
-              const Color(0xFF2E7D32), // Green theme for privacy/security
+          seedColor: const Color(
+            0xFF2E7D32,
+          ), // Green theme for privacy/security
           brightness: Brightness.light,
         ),
         useMaterial3: true,
-        appBarTheme: const AppBarTheme(
-          centerTitle: true,
-          elevation: 0,
-        ),
+        appBarTheme: const AppBarTheme(centerTitle: true, elevation: 0),
         cardTheme: CardThemeData(
           elevation: 2,
           shape: RoundedRectangleBorder(
@@ -73,11 +70,11 @@ class OCRixApp extends ConsumerWidget {
           ),
         ),
         inputDecorationTheme: InputDecorationTheme(
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(8),
+          border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
+          contentPadding: const EdgeInsets.symmetric(
+            horizontal: 16,
+            vertical: 12,
           ),
-          contentPadding:
-              const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
         ),
       ),
       darkTheme: ThemeData(
@@ -86,10 +83,7 @@ class OCRixApp extends ConsumerWidget {
           brightness: Brightness.dark,
         ),
         useMaterial3: true,
-        appBarTheme: const AppBarTheme(
-          centerTitle: true,
-          elevation: 0,
-        ),
+        appBarTheme: const AppBarTheme(centerTitle: true, elevation: 0),
         cardTheme: CardThemeData(
           elevation: 2,
           shape: RoundedRectangleBorder(
@@ -105,11 +99,11 @@ class OCRixApp extends ConsumerWidget {
           ),
         ),
         inputDecorationTheme: InputDecorationTheme(
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(8),
+          border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
+          contentPadding: const EdgeInsets.symmetric(
+            horizontal: 16,
+            vertical: 12,
           ),
-          contentPadding:
-              const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
         ),
       ),
       home: const AppInitializer(),
@@ -162,8 +156,9 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
       // Only check if we have a background time recorded (app was actually backgrounded)
       // and it was more than 1 second ago (to avoid checking on quick navigation)
       if (_lastBackgroundTime != null) {
-        final timeSinceBackground =
-            DateTime.now().difference(_lastBackgroundTime!);
+        final timeSinceBackground = DateTime.now().difference(
+          _lastBackgroundTime!,
+        );
         if (timeSinceBackground.inSeconds >= 1) {
           _checkBiometricOnResume();
         }
@@ -201,19 +196,22 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
     });
 
     try {
-      final biometricNotifier =
-          ref.read(biometricAuthNotifierProvider.notifier);
+      final biometricNotifier = ref.read(
+        biometricAuthNotifierProvider.notifier,
+      );
       final biometricService = ref.read(biometricAuthServiceProvider);
 
-      biometricService
-          .logInfo('App resumed - requesting biometric authentication');
+      biometricService.logInfo(
+        'App resumed - requesting biometric authentication',
+      );
       final authenticated = await biometricNotifier.authenticate(
         reason: 'Use your fingerprint to continue',
       );
 
       if (authenticated) {
-        biometricService
-            .logInfo('Biometric authentication successful on app resume');
+        biometricService.logInfo(
+          'Biometric authentication successful on app resume',
+        );
         setState(() {
           _isAuthenticated = true;
           _isCheckingBiometric = false;
@@ -221,7 +219,8 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
       } else {
         // Biometric failed or cancelled - show login screen with Google Sign-In as backup
         biometricService.logWarning(
-            'Biometric authentication failed or cancelled on app resume - showing login screen');
+          'Biometric authentication failed or cancelled on app resume - showing login screen',
+        );
         setState(() {
           _isAuthenticated = false;
           _isCheckingBiometric = false;
@@ -232,7 +231,10 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
       // On error, show login screen with Google Sign-In as backup
       final biometricService = ref.read(biometricAuthServiceProvider);
       biometricService.logError(
-          'Error during biometric check on app resume', e, stackTrace);
+        'Error during biometric check on app resume',
+        e,
+        stackTrace,
+      );
       setState(() {
         _isAuthenticated = false;
         _isCheckingBiometric = false;
@@ -268,36 +270,31 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
 
       // Inject troubleshooting logger into all services
       if (databaseService is DatabaseService) {
-        (databaseService as DatabaseService)
-            .setTroubleshootingLogger(troubleshootingLogger);
+        (databaseService).setTroubleshootingLogger(troubleshootingLogger);
       }
 
       if (encryptionService is EncryptionService) {
-        (encryptionService as EncryptionService)
-            .setTroubleshootingLogger(troubleshootingLogger);
+        (encryptionService).setTroubleshootingLogger(troubleshootingLogger);
       }
 
       if (ocrService is OCRService) {
-        (ocrService as OCRService)
-            .setTroubleshootingLogger(troubleshootingLogger);
+        (ocrService).setTroubleshootingLogger(troubleshootingLogger);
       }
 
-      if (cameraService is CameraService) {
-        (cameraService as CameraService)
-            .setTroubleshootingLogger(troubleshootingLogger);
-      }
+      (cameraService).setTroubleshootingLogger(troubleshootingLogger);
 
       if (storageService is StorageProviderService) {
-        (storageService as StorageProviderService)
-            .setTroubleshootingLogger(troubleshootingLogger);
+        (storageService).setTroubleshootingLogger(troubleshootingLogger);
       }
 
       // Initialize error handler
       ErrorHandler.initialize(troubleshootingLogger);
 
       // Log app initialization start
-      troubleshootingLogger.info('App initialization started',
-          tag: 'AppInitializer');
+      troubleshootingLogger.info(
+        'App initialization started',
+        tag: 'AppInitializer',
+      );
 
       // Initialize audit logging service (needed for DB logging)
       await auditLoggingService.initialize();
@@ -305,8 +302,7 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
       // Set audit logging service in database service for COMPULSORY logging
       // Cast to concrete type to access setAuditLoggingService
       if (databaseService is DatabaseService) {
-        (databaseService as DatabaseService)
-            .setAuditLoggingService(auditLoggingService);
+        (databaseService).setAuditLoggingService(auditLoggingService);
       }
 
       // Get current user ID for audit logging
@@ -319,8 +315,7 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
 
         // Set user ID in database for SQLite triggers
         if (databaseService is DatabaseService) {
-          await (databaseService as DatabaseService)
-              .setCurrentUserIdForTriggers(userId);
+          await (databaseService).setCurrentUserIdForTriggers(userId);
         }
       }
 
@@ -328,26 +323,36 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
 
       // Initialize critical services (must succeed)
       await databaseService.initialize();
-      troubleshootingLogger.info('Database service initialized',
-          tag: 'AppInitializer');
+      troubleshootingLogger.info(
+        'Database service initialized',
+        tag: 'AppInitializer',
+      );
 
       await encryptionService.initialize();
-      troubleshootingLogger.info('Encryption service initialized',
-          tag: 'AppInitializer');
+      troubleshootingLogger.info(
+        'Encryption service initialized',
+        tag: 'AppInitializer',
+      );
 
       await ocrService.initialize();
-      troubleshootingLogger.info('OCR service initialized',
-          tag: 'AppInitializer');
+      troubleshootingLogger.info(
+        'OCR service initialized',
+        tag: 'AppInitializer',
+      );
 
       await storageService.initialize();
-      troubleshootingLogger.info('Storage service initialized',
-          tag: 'AppInitializer');
+      troubleshootingLogger.info(
+        'Storage service initialized',
+        tag: 'AppInitializer',
+      );
 
       // Camera service is optional (may fail in CI/test environments)
       try {
         await cameraService.initialize();
-        troubleshootingLogger.info('Camera service initialized',
-            tag: 'AppInitializer');
+        troubleshootingLogger.info(
+          'Camera service initialized',
+          tag: 'AppInitializer',
+        );
       } catch (e) {
         // Log but don't fail app initialization if camera is unavailable
         // Camera features will be disabled, but app can still function
@@ -360,8 +365,10 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
         debugPrint('Warning: Camera service initialization failed: $e');
       }
 
-      troubleshootingLogger.info('App initialization completed successfully',
-          tag: 'AppInitializer');
+      troubleshootingLogger.info(
+        'App initialization completed successfully',
+        tag: 'AppInitializer',
+      );
 
       if (mounted) {
         setState(() {
@@ -410,11 +417,7 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              const Icon(
-                Icons.error_outline,
-                size: 64,
-                color: Colors.red,
-              ),
+              const Icon(Icons.error_outline, size: 64, color: Colors.red),
               const SizedBox(height: 16),
               Text(
                 'Failed to initialize app',
@@ -444,19 +447,14 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
       );
     }
 
-    // Show login screen if not signed in
-    if (!isSignedIn) {
-      // Reset authentication state when signed out
-      _isAuthenticated = false;
-      return const LoginScreen();
-    }
-
     if (!_isInitialized) {
       return const SplashScreen();
     }
 
-    // If biometric is enabled and user hasn't authenticated yet, show login
-    if (biometricState.isEnabled &&
+    // Only check biometric if user is signed in
+    // (Biometric is optional security layer, not required for app access)
+    if (isSignedIn &&
+        biometricState.isEnabled &&
         biometricState.isAvailable &&
         !_isAuthenticated &&
         !_isCheckingBiometric) {
@@ -470,6 +468,8 @@ class _AppInitializerState extends ConsumerState<AppInitializer>
       return const SplashScreen();
     }
 
+    // App can be used without Google Sign-In
+    // Sign-in is only required for cloud sync/export features
     return const HomeScreen();
   }
 }
