@@ -208,35 +208,35 @@ class Document extends Equatable {
 
   @override
   List<Object?> get props => [
-        id,
-        title,
-        imageData,
-        thumbnailData,
-        imageFormat,
-        imageSize,
-        imageWidth,
-        imageHeight,
-        imagePath,
-        extractedText,
-        type,
-        scanDate,
-        tags,
-        metadata,
-        storageProvider,
-        isEncrypted,
-        confidenceScore,
-        detectedLanguage,
-        deviceInfo,
-        notes,
-        location,
-        createdAt,
-        updatedAt,
-        isSynced,
-        cloudId,
-        lastSyncedAt,
-        isMultiPage,
-        pageCount,
-      ];
+    id,
+    title,
+    imageData,
+    thumbnailData,
+    imageFormat,
+    imageSize,
+    imageWidth,
+    imageHeight,
+    imagePath,
+    extractedText,
+    type,
+    scanDate,
+    tags,
+    metadata,
+    storageProvider,
+    isEncrypted,
+    confidenceScore,
+    detectedLanguage,
+    deviceInfo,
+    notes,
+    location,
+    createdAt,
+    updatedAt,
+    isSynced,
+    cloudId,
+    lastSyncedAt,
+    isMultiPage,
+    pageCount,
+  ];
 }
 
 enum DocumentType {
