@@ -112,22 +112,22 @@ class UserSettings extends Equatable {
 
   @override
   List<Object?> get props => [
-        metadataStorageProvider,
-        fileStorageProvider,
-        autoSync,
-        syncIntervalMinutes,
-        biometricAuth,
-        encryptionEnabled,
-        defaultDocumentType,
-        defaultTags,
-        privacyAuditEnabled,
-        language,
-        theme,
-        notificationsEnabled,
-        autoCategorization,
-        ocrConfidenceThreshold,
-        backupEnabled,
-        lastBackupAt,
-        customSettings,
-      ];
+    metadataStorageProvider,
+    fileStorageProvider,
+    autoSync,
+    syncIntervalMinutes,
+    biometricAuth,
+    encryptionEnabled,
+    defaultDocumentType,
+    defaultTags,
+    privacyAuditEnabled,
+    language,
+    theme,
+    notificationsEnabled,
+    autoCategorization,
+    ocrConfidenceThreshold,
+    backupEnabled,
+    lastBackupAt,
+    customSettings,
+  ];
 }
