@@ -23,8 +23,6 @@ import 'audit_provider.dart';
 import 'troubleshooting_logger_provider.dart';
 import '../core/interfaces/troubleshooting_logger_interface.dart';
 import '../models/document_page.dart';
-import 'package:path/path.dart' as path;
-import 'package:path_provider/path_provider.dart';
 
 // Service providers - using interfaces for dependency inversion
 final databaseServiceProvider = Provider<IDatabaseService>((ref) {
@@ -55,14 +53,18 @@ final documentListProvider = FutureProvider<List<Document>>((ref) async {
   return await databaseService.getAllDocuments();
 });
 
-final documentProvider =
-    FutureProvider.family<Document?, String>((ref, documentId) async {
+final documentProvider = FutureProvider.family<Document?, String>((
+  ref,
+  documentId,
+) async {
   final databaseService = ref.read(databaseServiceProvider);
   return await databaseService.getDocument(documentId);
 });
 
-final documentSearchProvider =
-    FutureProvider.family<List<Document>, String>((ref, query) async {
+final documentSearchProvider = FutureProvider.family<List<Document>, String>((
+  ref,
+  query,
+) async {
   final databaseService = ref.read(databaseServiceProvider);
   if (query.isEmpty) {
     return await databaseService.getAllDocuments();
@@ -72,9 +74,9 @@ final documentSearchProvider =
 
 final documentByTypeProvider =
     FutureProvider.family<List<Document>, DocumentType>((ref, type) async {
-  final databaseService = ref.read(databaseServiceProvider);
-  return await databaseService.getAllDocuments(type: type);
-});
+      final databaseService = ref.read(databaseServiceProvider);
+      return await databaseService.getAllDocuments(type: type);
+    });
 
 class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
   final IDatabaseService _databaseService;
@@ -96,11 +98,11 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
     IImageProcessingService? imageProcessingService,
     AuditLoggingService? auditLoggingService,
     ITroubleshootingLogger? troubleshootingLogger,
-  })  : _imageProcessingService =
-            imageProcessingService ?? ImageProcessingService(),
-        _auditLoggingService = auditLoggingService,
-        _troubleshootingLogger = troubleshootingLogger,
-        super(const AsyncValue.loading()) {
+  }) : _imageProcessingService =
+           imageProcessingService ?? ImageProcessingService(),
+       _auditLoggingService = auditLoggingService,
+       _troubleshootingLogger = troubleshootingLogger,
+       super(const AsyncValue.loading()) {
     _loadDocuments();
   }
 
@@ -122,7 +124,7 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
         _currentSearchQuery = searchQuery ?? _currentSearchQuery;
       }
 
-      final limit = AppConfig.documentsPerPage;
+      const limit = AppConfig.documentsPerPage;
       final offset = _currentPage * limit;
 
       final documents = await _databaseService.getAllDocuments(
@@ -206,10 +208,8 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
       }
 
       // Process image for optimal storage using ImageProcessingService
-      final processedResult =
-          await _imageProcessingService.processImageForStorage(
-        imageBytes.toList(),
-      );
+      final processedResult = await _imageProcessingService
+          .processImageForStorage(imageBytes.toList());
 
       // Extract text using OCR from original image
       final ocrResult = await _ocrService.extractTextFromImage(imagePath);
@@ -309,14 +309,13 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
       }
 
       // Combine extracted text from all pages
-      final combinedText =
-          pages.map((p) => p.extractedText).join('\n\n--- Page Break ---\n\n');
+      final combinedText = pages
+          .map((p) => p.extractedText)
+          .join('\n\n--- Page Break ---\n\n');
 
       // Use the first page's confidence score as overall confidence
-      final avgConfidence = pages.fold<double>(
-            0.0,
-            (sum, page) => sum + page.confidenceScore,
-          ) /
+      final avgConfidence =
+          pages.fold<double>(0.0, (sum, page) => sum + page.confidenceScore) /
           pages.length;
 
       // Categorize document based on combined text
@@ -352,9 +351,7 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
 
       // Update pages with document ID and save them
       for (int i = 0; i < pages.length; i++) {
-        final page = pages[i].copyWith(
-          documentId: documentId,
-        );
+        final page = pages[i].copyWith(documentId: documentId);
         await _databaseService.saveDocumentPage(page);
       }
 
@@ -394,9 +391,7 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
 
   Future<void> updateDocument(Document document) async {
     try {
-      final updatedDocument = document.copyWith(
-        updatedAt: DateTime.now(),
-      );
+      final updatedDocument = document.copyWith(updatedAt: DateTime.now());
 
       await _databaseService.updateDocument(updatedDocument);
 
@@ -456,9 +451,7 @@ class DocumentNotifier extends StateNotifier<AsyncValue<List<Document>>> {
       // Optimistic update: remove document from state directly
       if (state.hasValue) {
         final docs = state.value ?? [];
-        state = AsyncValue.data(
-          docs.where((d) => d.id != documentId).toList(),
-        );
+        state = AsyncValue.data(docs.where((d) => d.id != documentId).toList());
       } else {
         await refreshDocuments();
       }
@@ -528,25 +521,25 @@ final imageProcessingServiceProvider = Provider<IImageProcessingService>((ref) {
 
 final documentNotifierProvider =
     StateNotifierProvider<DocumentNotifier, AsyncValue<List<Document>>>((ref) {
-  final databaseService = ref.read(databaseServiceProvider);
-  final ocrService = ref.read(ocrServiceProvider);
-  final cameraService = ref.read(cameraServiceProvider);
-  final storageService = ref.read(storageProviderServiceProvider);
-  final imageProcessingService = ref.read(imageProcessingServiceProvider);
-  final auditLoggingService = ref.read(auditLoggingServiceProvider);
+      final databaseService = ref.read(databaseServiceProvider);
+      final ocrService = ref.read(ocrServiceProvider);
+      final cameraService = ref.read(cameraServiceProvider);
+      final storageService = ref.read(storageProviderServiceProvider);
+      final imageProcessingService = ref.read(imageProcessingServiceProvider);
+      final auditLoggingService = ref.read(auditLoggingServiceProvider);
 
-  final troubleshootingLogger = ref.read(troubleshootingLoggerProvider);
+      final troubleshootingLogger = ref.read(troubleshootingLoggerProvider);
 
-  return DocumentNotifier(
-    databaseService,
-    ocrService,
-    cameraService,
-    storageService,
-    imageProcessingService: imageProcessingService,
-    auditLoggingService: auditLoggingService,
-    troubleshootingLogger: troubleshootingLogger,
-  );
-});
+      return DocumentNotifier(
+        databaseService,
+        ocrService,
+        cameraService,
+        storageService,
+        imageProcessingService: imageProcessingService,
+        auditLoggingService: auditLoggingService,
+        troubleshootingLogger: troubleshootingLogger,
+      );
+    });
 
 class DocumentDetailNotifier extends StateNotifier<AsyncValue<Document?>> {
   final IDatabaseService _databaseService;
@@ -555,8 +548,8 @@ class DocumentDetailNotifier extends StateNotifier<AsyncValue<Document?>> {
   DocumentDetailNotifier(
     this._databaseService, {
     ITroubleshootingLogger? troubleshootingLogger,
-  })  : _troubleshootingLogger = troubleshootingLogger,
-        super(const AsyncValue.loading());
+  }) : _troubleshootingLogger = troubleshootingLogger,
+       super(const AsyncValue.loading());
 
   Future<void> loadDocument(String documentId) async {
     try {
@@ -576,9 +569,7 @@ class DocumentDetailNotifier extends StateNotifier<AsyncValue<Document?>> {
 
   Future<void> updateDocument(Document document) async {
     try {
-      final updatedDocument = document.copyWith(
-        updatedAt: DateTime.now(),
-      );
+      final updatedDocument = document.copyWith(updatedAt: DateTime.now());
 
       await _databaseService.updateDocument(updatedDocument);
       state = AsyncValue.data(updatedDocument);
@@ -619,15 +610,19 @@ class DocumentDetailNotifier extends StateNotifier<AsyncValue<Document?>> {
   }
 }
 
-final documentDetailNotifierProvider = StateNotifierProvider.family<
-    DocumentDetailNotifier, AsyncValue<Document?>, String>((ref, documentId) {
-  final databaseService = ref.read(databaseServiceProvider);
-  final troubleshootingLogger = ref.read(troubleshootingLoggerProvider);
-  return DocumentDetailNotifier(
-    databaseService,
-    troubleshootingLogger: troubleshootingLogger,
-  );
-});
+final documentDetailNotifierProvider =
+    StateNotifierProvider.family<
+      DocumentDetailNotifier,
+      AsyncValue<Document?>,
+      String
+    >((ref, documentId) {
+      final databaseService = ref.read(databaseServiceProvider);
+      final troubleshootingLogger = ref.read(troubleshootingLoggerProvider);
+      return DocumentDetailNotifier(
+        databaseService,
+        troubleshootingLogger: troubleshootingLogger,
+      );
+    });
 
 class ScannerNotifier extends StateNotifier<ScannerState> {
   final ICameraService _cameraService;
@@ -638,8 +633,8 @@ class ScannerNotifier extends StateNotifier<ScannerState> {
     this._cameraService,
     this._ocrService, {
     ITroubleshootingLogger? troubleshootingLogger,
-  })  : _troubleshootingLogger = troubleshootingLogger,
-        super(const ScannerState.initial());
+  }) : _troubleshootingLogger = troubleshootingLogger,
+       super(const ScannerState.initial());
 
   Future<void> initializeCamera() async {
     try {
@@ -664,10 +659,7 @@ class ScannerNotifier extends StateNotifier<ScannerState> {
         tag: 'ScannerNotifier',
         error: e,
       );
-      state = state.copyWith(
-        isLoading: false,
-        error: e.toString(),
-      );
+      state = state.copyWith(isLoading: false, error: e.toString());
     }
   }
 
@@ -694,10 +686,7 @@ class ScannerNotifier extends StateNotifier<ScannerState> {
         tag: 'ScannerNotifier',
         error: e,
       );
-      state = state.copyWith(
-        isCapturing: false,
-        error: e.toString(),
-      );
+      state = state.copyWith(isCapturing: false, error: e.toString());
       return null;
     }
   }
@@ -725,10 +714,7 @@ class ScannerNotifier extends StateNotifier<ScannerState> {
         tag: 'ScannerNotifier',
         error: e,
       );
-      state = state.copyWith(
-        isProcessing: false,
-        error: e.toString(),
-      );
+      state = state.copyWith(isProcessing: false, error: e.toString());
       return null;
     }
   }
@@ -748,15 +734,15 @@ class ScannerNotifier extends StateNotifier<ScannerState> {
 
 final scannerNotifierProvider =
     StateNotifierProvider<ScannerNotifier, ScannerState>((ref) {
-  final cameraService = ref.read(cameraServiceProvider);
-  final ocrService = ref.read(ocrServiceProvider);
-  final troubleshootingLogger = ref.read(troubleshootingLoggerProvider);
-  return ScannerNotifier(
-    cameraService,
-    ocrService,
-    troubleshootingLogger: troubleshootingLogger,
-  );
-});
+      final cameraService = ref.read(cameraServiceProvider);
+      final ocrService = ref.read(ocrServiceProvider);
+      final troubleshootingLogger = ref.read(troubleshootingLoggerProvider);
+      return ScannerNotifier(
+        cameraService,
+        ocrService,
+        troubleshootingLogger: troubleshootingLogger,
+      );
+    });
 
 class ScannerState {
   final bool isLoading;
