@@ -82,7 +82,7 @@ class EncryptionService extends BaseService implements IEncryptionService {
       }
 
       if (_encrypter == null || _key == null) {
-        throw EncryptionException('Encryption not initialized');
+        throw const EncryptionException('Encryption not initialized');
       }
 
       // Generate a unique IV for this encryption operation
@@ -118,7 +118,7 @@ class EncryptionService extends BaseService implements IEncryptionService {
       }
 
       if (_encrypter == null || _key == null) {
-        throw EncryptionException('Encryption not initialized');
+        throw const EncryptionException('Encryption not initialized');
       }
 
       // Decode base64 to get [IV (16 bytes)][Encrypted Data]
@@ -126,7 +126,8 @@ class EncryptionService extends BaseService implements IEncryptionService {
 
       if (encryptedBytes.length < 16) {
         throw EncryptionException(
-            'Invalid encrypted text: too short (${encryptedBytes.length} bytes)');
+          'Invalid encrypted text: too short (${encryptedBytes.length} bytes)',
+        );
       }
 
       // Extract IV from first 16 bytes
@@ -162,7 +163,7 @@ class EncryptionService extends BaseService implements IEncryptionService {
       }
 
       if (_encrypter == null || _key == null) {
-        throw EncryptionException('Encryption not initialized');
+        throw const EncryptionException('Encryption not initialized');
       }
 
       // Generate a unique IV for this encryption operation
@@ -180,7 +181,8 @@ class EncryptionService extends BaseService implements IEncryptionService {
       result.setRange(16, result.length, encrypted.bytes);
 
       logInfo(
-          'Bytes encrypted successfully: ${bytes.length} -> ${result.length} (IV: 16 bytes + encrypted: ${encrypted.bytes.length} bytes)');
+        'Bytes encrypted successfully: ${bytes.length} -> ${result.length} (IV: 16 bytes + encrypted: ${encrypted.bytes.length} bytes)',
+      );
       return result.toList();
     } catch (e) {
       logError('Failed to encrypt bytes', e);
@@ -202,13 +204,14 @@ class EncryptionService extends BaseService implements IEncryptionService {
       }
 
       if (_encrypter == null || _key == null) {
-        throw EncryptionException('Encryption not initialized');
+        throw const EncryptionException('Encryption not initialized');
       }
 
       // Encrypted data format: [IV (16 bytes)][Encrypted Data]
       if (encryptedBytes.length < 16) {
         throw EncryptionException(
-            'Invalid encrypted data: too short (${encryptedBytes.length} bytes)');
+          'Invalid encrypted data: too short (${encryptedBytes.length} bytes)',
+        );
       }
 
       // Extract IV from first 16 bytes
@@ -223,7 +226,8 @@ class EncryptionService extends BaseService implements IEncryptionService {
       final decrypted = encrypter.decryptBytes(encrypted, iv: extractedIV);
 
       logInfo(
-          'Bytes decrypted successfully: ${encryptedBytes.length} (IV: 16 bytes + encrypted: ${encryptedData.length} bytes) -> ${decrypted.length}');
+        'Bytes decrypted successfully: ${encryptedBytes.length} (IV: 16 bytes + encrypted: ${encryptedData.length} bytes) -> ${decrypted.length}',
+      );
       return decrypted.toList();
     } catch (e) {
       logError('Failed to decrypt bytes', e);
@@ -273,7 +277,8 @@ class EncryptionService extends BaseService implements IEncryptionService {
       final encryptedFile = File(encryptedFilePath);
       if (!await encryptedFile.exists()) {
         throw EncryptionException(
-            'Encrypted file does not exist: $encryptedFilePath');
+          'Encrypted file does not exist: $encryptedFilePath',
+        );
       }
 
       final encryptedBytes = await encryptedFile.readAsBytes();
@@ -318,10 +323,8 @@ class EncryptionService extends BaseService implements IEncryptionService {
       // Authenticate
       final isAuthenticated = await _localAuth.authenticate(
         localizedReason: reason ?? 'Authenticate to access encrypted documents',
-        options: const AuthenticationOptions(
-          biometricOnly: true,
-          stickyAuth: true,
-        ),
+        biometricOnly: true,
+        persistAcrossBackgrounding: true,
       );
 
       if (isAuthenticated) {
@@ -415,10 +418,8 @@ class EncryptionService extends BaseService implements IEncryptionService {
   /// Derive an encryption key from a password using PBKDF2
   /// Returns: [Key, Salt] - both are needed for encryption/decryption
   List<Uint8List> _deriveKeyFromPassword(String password, Uint8List? salt) {
-    // Use provided salt or generate new one
-    final keySalt = salt ??
-        Uint8List.fromList(List<int>.generate(
-            32, (i) => DateTime.now().millisecondsSinceEpoch % 256 + i));
+    // Use provided salt or generate cryptographically secure random salt
+    final keySalt = salt ?? IV.fromSecureRandom(32).bytes;
 
     // PBKDF2 parameters
     const iterations = 100000; // High iteration count for security
