@@ -13,6 +13,7 @@ import '../core/interfaces/encryption_service_interface.dart';
 import '../core/base/base_service.dart';
 import '../core/exceptions/app_exceptions.dart';
 import 'encryption_service.dart';
+import 'database_service.dart';
 
 abstract class StorageProviderInterface {
   Future<bool> initialize();
@@ -31,7 +32,7 @@ class LocalStorageProvider implements StorageProviderInterface {
   final String _basePath;
 
   LocalStorageProvider({String? basePath})
-      : _basePath = basePath ?? 'documents';
+    : _basePath = basePath ?? 'documents';
 
   @override
   Future<bool> initialize() async {
@@ -133,8 +134,10 @@ class LocalStorageProvider implements StorageProviderInterface {
       final files = <String>[];
       await for (final entity in storageDir.list(recursive: true)) {
         if (entity is File) {
-          final relativePath =
-              path.relative(entity.path, from: storageDir.path);
+          final relativePath = path.relative(
+            entity.path,
+            from: storageDir.path,
+          );
           if (prefix == null || relativePath.startsWith(prefix)) {
             files.add(relativePath);
           }
@@ -164,15 +167,15 @@ class GoogleDriveStorageProvider implements StorageProviderInterface {
   // Note: This is not a service, so it doesn't extend BaseService
   // Using Logger directly for now
   final logger = Logger();
-  final GoogleSignIn _googleSignIn = GoogleSignIn(
-    scopes: [
-      drive.DriveApi.driveFileScope,
-      'https://www.googleapis.com/auth/drive.appdata', // Required for appDataFolder access
-    ],
-  );
+  final GoogleSignIn _googleSignIn = GoogleSignIn.instance;
+  final List<String> _scopes = [
+    drive.DriveApi.driveFileScope,
+    'https://www.googleapis.com/auth/drive.appdata', // Required for appDataFolder access
+  ];
 
   drive.DriveApi? _driveApi;
   bool _isInitialized = false;
+  GoogleSignInAccount? _currentUser;
 
   // Expose driveApi for advanced operations (like getting file metadata)
   drive.DriveApi? get driveApi => _driveApi;
@@ -180,34 +183,92 @@ class GoogleDriveStorageProvider implements StorageProviderInterface {
   @override
   Future<bool> initialize() async {
     try {
-      final account = await _googleSignIn.signIn();
-      if (account == null) {
-        logger.w('Google Sign-In cancelled');
+      // Initialize GoogleSignIn if needed (this is safe to call multiple times)
+      try {
+        await _googleSignIn.initialize(
+          serverClientId:
+              '340615948692-mqara4jiq5l52pp7027cm16s9eojc5vg.apps.googleusercontent.com',
+        );
+      } catch (e) {
+        // Already initialized, ignore
+      }
+
+      // Authenticate user for Drive access
+      // First try lightweight authentication
+      GoogleSignInAccount? user;
+      final result = _googleSignIn.attemptLightweightAuthentication();
+      if (result is Future<GoogleSignInAccount?>) {
+        user = await result;
+      }
+
+      // If lightweight auth didn't work, do full authentication
+      if (user == null) {
+        user = await _googleSignIn.authenticate(scopeHint: _scopes);
+      }
+
+      if (user == null) {
+        logger.w('Google Sign-In cancelled or failed');
         return false;
       }
 
-      final auth = await account.authentication;
-      if (auth.accessToken == null) {
-        logger.e('Failed to get access token');
-        return false;
-      }
+      _currentUser = user;
 
-      final authClient = authenticatedClient(
-        http.Client(),
-        AccessCredentials(
-          AccessToken('Bearer', auth.accessToken!,
-              DateTime.now().toUtc().add(const Duration(hours: 1))),
-          auth.idToken,
-          [
-            drive.DriveApi.driveFileScope,
-            'https://www.googleapis.com/auth/drive.appdata', // Required for appDataFolder access
-          ],
-        ),
+      // Get access token for Drive API using authorizationClient
+      final clientAuth = await user.authorizationClient.authorizationForScopes(
+        _scopes,
       );
 
-      _driveApi = drive.DriveApi(authClient);
-      _isInitialized = true;
+      if (clientAuth == null) {
+        // Need to authorize scopes
+        logger.i('Requesting authorization for Drive scopes');
+        final newAuth = await user.authorizationClient.authorizeScopes(_scopes);
+        if (newAuth == null) {
+          logger.e('Failed to authorize Drive scopes');
+          return false;
+        }
 
+        // Get the access token after authorization
+        final finalAuth = await user.authorizationClient.authorizationForScopes(
+          _scopes,
+        );
+        if (finalAuth == null) {
+          logger.e('Failed to get access token after authorization');
+          return false;
+        }
+
+        final authClient = authenticatedClient(
+          http.Client(),
+          AccessCredentials(
+            AccessToken(
+              'Bearer',
+              finalAuth.accessToken,
+              DateTime.now().toUtc().add(const Duration(hours: 1)),
+            ),
+            null,
+            _scopes,
+          ),
+        );
+
+        _driveApi = drive.DriveApi(authClient);
+      } else {
+        // Already have authorization
+        final authClient = authenticatedClient(
+          http.Client(),
+          AccessCredentials(
+            AccessToken(
+              'Bearer',
+              clientAuth.accessToken,
+              DateTime.now().toUtc().add(const Duration(hours: 1)),
+            ),
+            null,
+            _scopes,
+          ),
+        );
+
+        _driveApi = drive.DriveApi(authClient);
+      }
+
+      _isInitialized = true;
       logger.i('Google Drive storage provider initialized');
       return true;
     } catch (e) {
@@ -255,10 +316,12 @@ class GoogleDriveStorageProvider implements StorageProviderInterface {
         throw Exception('Google Drive not initialized');
       }
 
-      final media = await _driveApi!.files.get(
-        remotePath,
-        downloadOptions: drive.DownloadOptions.fullMedia,
-      ) as drive.Media;
+      final media =
+          await _driveApi!.files.get(
+                remotePath,
+                downloadOptions: drive.DownloadOptions.fullMedia,
+              )
+              as drive.Media;
 
       final bytes = <int>[];
       await for (final chunk in media.stream) {
@@ -304,7 +367,8 @@ class GoogleDriveStorageProvider implements StorageProviderInterface {
         spaces: 'appDataFolder',
       );
 
-      final fileIds = files.files
+      final fileIds =
+          files.files
               ?.map((file) => file.id ?? '')
               .where((id) => id.isNotEmpty)
               .toList() ??
@@ -321,7 +385,7 @@ class GoogleDriveStorageProvider implements StorageProviderInterface {
   @override
   Future<bool> isConnected() async {
     try {
-      return _isInitialized && await _googleSignIn.isSignedIn();
+      return _isInitialized && _currentUser != null;
     } catch (e) {
       logger.e('Failed to check Google Drive connection: $e');
       return false;
@@ -334,6 +398,7 @@ class GoogleDriveStorageProvider implements StorageProviderInterface {
       await _googleSignIn.signOut();
       _driveApi = null;
       _isInitialized = false;
+      _currentUser = null;
       logger.i('Disconnected from Google Drive');
     } catch (e) {
       logger.e('Failed to disconnect from Google Drive: $e');
@@ -551,14 +616,122 @@ class StorageProviderService extends BaseService
 
   @override
   Future<void> syncToCloud() async {
-    // TODO: Implement cloud sync
-    logInfo('Sync to cloud - not yet implemented');
+    try {
+      logInfo('Starting cloud sync (upload)');
+
+      // Check if cloud provider is configured
+      if (_currentProvider == StorageProviderType.local) {
+        logWarning('Cannot sync to cloud: No cloud provider configured');
+        return;
+      }
+
+      // Get database service to fetch unsynced documents
+      final dbService = DatabaseService();
+      final documents = await dbService.getAllDocuments();
+
+      // Filter unsynced documents
+      final unsyncedDocs = documents.where((doc) => !doc.isSynced).toList();
+
+      if (unsyncedDocs.isEmpty) {
+        logInfo('No documents to sync');
+        return;
+      }
+
+      logInfo('Found ${unsyncedDocs.length} documents to sync');
+
+      // Upload each unsynced document
+      int syncedCount = 0;
+      for (final doc in unsyncedDocs) {
+        try {
+          await uploadDocument(doc);
+
+          // Mark as synced in database
+          final updatedDoc = doc.copyWith(
+            updatedAt: DateTime.now(),
+            isSynced: true,
+          );
+
+          await dbService.updateDocument(updatedDoc);
+          syncedCount++;
+
+          logInfo('Synced document: ${doc.title}');
+        } catch (e) {
+          logError('Failed to sync document ${doc.title}', e);
+          // Continue with next document
+        }
+      }
+
+      logInfo(
+        'Cloud sync completed: $syncedCount/${unsyncedDocs.length} documents synced',
+      );
+    } catch (e) {
+      logError('Failed to sync to cloud', e);
+      throw StorageException(
+        'Failed to sync to cloud: ${e.toString()}',
+        originalError: e,
+      );
+    }
   }
 
   @override
   Future<void> syncFromCloud() async {
-    // TODO: Implement cloud sync
-    logInfo('Sync from cloud - not yet implemented');
+    try {
+      logInfo('Starting cloud sync (download)');
+
+      // Check if cloud provider is configured
+      if (_currentProvider == StorageProviderType.local) {
+        logWarning('Cannot sync from cloud: No cloud provider configured');
+        return;
+      }
+
+      // Get list of files from cloud
+      final provider = await getProvider(_currentProvider!);
+      final cloudFiles = await provider.listFiles(null);
+
+      if (cloudFiles.isEmpty) {
+        logInfo('No files found in cloud storage');
+        return;
+      }
+
+      logInfo('Found ${cloudFiles.length} files in cloud');
+
+      // Get database service to check existing documents
+      final dbService = DatabaseService();
+      final localDocs = await dbService.getAllDocuments();
+      final localFileNames = localDocs
+          .where((d) => d.imagePath != null)
+          .map((d) => path.basename(d.imagePath!))
+          .toSet();
+
+      // Download files that don't exist locally
+      int downloadedCount = 0;
+      final directory = await getApplicationDocumentsDirectory();
+
+      for (final cloudFile in cloudFiles) {
+        final fileName = path.basename(cloudFile);
+
+        if (!localFileNames.contains(fileName)) {
+          try {
+            final localPath = path.join(directory.path, 'downloads', fileName);
+            await provider.downloadFile(cloudFile, localPath);
+            downloadedCount++;
+
+            logInfo('Downloaded file from cloud: $fileName');
+          } catch (e) {
+            logError('Failed to download file $fileName', e);
+            // Continue with next file
+          }
+        }
+      }
+
+      logInfo('Cloud sync completed: $downloadedCount new files downloaded');
+    } catch (e) {
+      logError('Failed to sync from cloud', e);
+      throw StorageException(
+        'Failed to sync from cloud: ${e.toString()}',
+        originalError: e,
+      );
+    }
   }
 
   // Backward compatibility methods
