@@ -21,8 +21,9 @@ class ErrorHandler {
         metadata: {
           'library': details.library,
           'context': details.context?.toString(),
-          'informationCollector':
-              details.informationCollector?.call().toString(),
+          'informationCollector': details.informationCollector
+              ?.call()
+              .toString(),
         },
       );
 
