@@ -34,8 +34,9 @@ class AuditNavigationObserver extends RouteObserver<PageRoute<dynamic>> {
     if (auditService == null) return;
 
     final toScreen = _getScreenName(toRoute);
-    final fromScreen =
-        fromRoute != null ? _getScreenName(fromRoute) : 'unknown';
+    final fromScreen = fromRoute != null
+        ? _getScreenName(fromRoute)
+        : 'unknown';
 
     // Log navigation (VERBOSE level)
     auditService.logNavigation(
@@ -57,8 +58,9 @@ class AuditNavigationObserver extends RouteObserver<PageRoute<dynamic>> {
     final routeString = route.toString();
     if (routeString.contains('HomeScreen')) return 'HomeScreen';
     if (routeString.contains('DocumentListScreen')) return 'DocumentListScreen';
-    if (routeString.contains('DocumentDetailScreen'))
+    if (routeString.contains('DocumentDetailScreen')) {
       return 'DocumentDetailScreen';
+    }
     if (routeString.contains('ScannerScreen')) return 'ScannerScreen';
     if (routeString.contains('SettingsScreen')) return 'SettingsScreen';
     if (routeString.contains('LoginScreen')) return 'LoginScreen';
